Move most of main package into common/obfs4proxy
To enable easy embedding of obfs4proxy into any Go program.
